requirements.txt
fastapi
uvicorn
psycopg2
click
python-dotenv
